Group 1 Final Project
THis project a challenge on Zindi to predict the estimated time of delivery of orders, from the point of driver pickup to the point of arrival at final destination. For this project, a training dataset from a subset of about 20,000 orders with bikes (“boda-boda”) in Nairobi is provided. This data is also anonymized. There are also additional GIS dataset that can be leveraged for this project. 
A comprehensive exploratory data analysis (EDA) was prepared using the training dataset. The data was checked for missing values and cleaned. New features were engineered on selected variables A multiple linear regression using Lasso, Ridge and Elastic Net was modelled and their MSE, RMSE compared. The MSE of  data with new feature and the ones without the features were also took into consideration.
This was worked on by 
Foster Twumasi Ankrah
Obed Baah Amoah
McSampson Keteku